src/backend/requirements.txt
# Core Dependencies
fastapi==0.103.1
uvicorn==0.23.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.20
alembic==1.12.0
pydantic==2.3.0
pydantic-settings==2.0.3
python-dotenv==1.0.0
openai==1.3.5

# Database
psycopg2-binary==2.9.7
SQLAlchemy-Utils==0.41.1

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx>=0.25.0
pytest-httpx>=0.25.0
pytest-mock==3.12.0

# Development
black==23.9.1
isort==5.12.0
mypy==1.5.1
pylint==3.0.2
pytest-watch==4.2.0

# Documentation
mkdocs==1.5.2
mkdocs-material==9.2.3

# Async
anyio==3.7.1

# Security
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
